chore(rpc): move EthRpcErrorCode to rpc-types
